func_tests/requirements3.txt
urllib3==1.22
pytest==3.0.7
pytest-timeout==1.2.1
pytest-cpp==0.4.5
pytz==2017.3
tzlocal==1.5.1
requests==2.18.4
Jinja2==2.10
netaddr==0.7.19
pytest-instafail==0.3.0
PyYAML==3.12
pathlib2==2.3.0
pylru==1.0.9
datadiff==2.0.0
pywinrm==0.3.0
xmltodict==0.11.0
cryptography==2.2.2
pysmb==1.1.22
python-dateutil==2.7.2
paramiko==2.4.1
click==6.7
Flask==0.12.2
pathlib2==2.3.0
Werkzeug==0.14.1
pyOpenSSL==18.0.0
typing==3.6.4
six==1.11.0
contextlib2==0.5.5
pycrypto==2.6.1
parse==1.8.4
monotonic==1.5
ffmpeg-python==0.1.16
ifaddr==0.1.4
opencv-python==3.4.2.17
portalocker==1.1.0
hachoir3==3.0a2
configparser==3.5.0
subprocess32==3.5.2
pytest-html==1.19.0
pytest-metadata==1.7.0
oyaml==0.6
